controller, model 분리 후 apiUtil, evnetUtil 구조 추가
