+++ src/model/apiUtil/StockInfoAPI.java
@@ -1,6 +1,87 @@
 package model.apiUtil;
 
+import java.util.Objects;
+
+import javax.swing.DefaultListModel;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+
+import model.StockDividendInfoModel;
+import model.StockInfoModel;
+import model.vo.StockInfoVO;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+
 // 주식 시세 정보를 호출하는 API
 public class StockInfoAPI {
+
+	private static final String STOCKINFO_API = "https://apis.data.go.kr/1160100/service/GetStockSecuritiesInfoService/getStockPriceInfo?serviceKey="
+			+ StockDeividendInfoAPI.apiKey + "&numOfRows=1&pageNo=1&resultType=json&itmsNm=";
+
+	// 통신 객체 (HTTP Client : HTTP 요청을 보내고 응답받는 객체)
+	private static final OkHttpClient client = new OkHttpClient();
+	// Gson 객체
+	private static Gson gson = new GsonBuilder().setPrettyPrinting().create();
+
+	// 주식 배당금 api에서 사용할 isin 코드를 반환하는 메서드
+	public static String getIsinCd(String itmsNm) {
+
+		Request request = new Request.Builder().url(STOCKINFO_API + itmsNm).build();
+		String isinCd = null;
+
+		try {
+			Response response = client.newCall(request).execute();
+			String json = Objects.requireNonNull(response.body()).string();
+			JsonObject jsonObj = gson.fromJson(json, JsonObject.class);
+			JsonArray itemArray = jsonObj.getAsJsonObject("response").getAsJsonObject("body").getAsJsonObject("items")
+					.getAsJsonArray("item");
+
+			isinCd = itemArray.get(0).getAsJsonObject().get("isinCd").getAsString();
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		return isinCd;
+
+	} // getIsinCd
+
+	// 주식 시세 정보에서 시가총액, isinCode, 시가, 종가, 고가, 저가 데이터를 담는 listModel 반환하는 메서드
+	public static DefaultListModel<StockInfoVO> getStockInfo(String itmsNm) {
+
+		Request request = new Request.Builder().url(STOCKINFO_API + itmsNm).build();
+
+		try {
+			Response response = client.newCall(request).execute();
+			String json = Objects.requireNonNull(response.body()).string();
+			JsonObject jsonObj = gson.fromJson(json, JsonObject.class);
+			JsonArray itemArray = jsonObj.getAsJsonObject("response").getAsJsonObject("body").getAsJsonObject("items")
+					.getAsJsonArray("item");
+
+			for (JsonElement ele : itemArray) {
+				JsonObject obj = ele.getAsJsonObject();
+
+				StockInfoVO stockInfoVO = new StockInfoVO(obj.get("mrktTotAmt").getAsString(), // 시가총액
+						obj.get("isinCd").getAsString(), // isinCode
+						obj.get("mkp").getAsString(), // 시가
+						obj.get("clpr").getAsString(), // 종가
+						obj.get("hipr").getAsString(), // 고가
+						obj.get("lopr").getAsString() // 저가
+				);
+
+				StockInfoModel.getStockInfoList().addElement(stockInfoVO);
+			}
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+
+		return StockInfoModel.getStockInfoList();
+
+	} // getStockInfo
 
 }
